Add separate view for CF handling - WIP.
CFs will be managed in a separate view in the future. This
commit prepares the skeleton for the new view and implements
a first version for managing CF destination sets.

The current CF handling is not touched and the new stuff
won't show up unless you add the "cf_test" preference
manually to the db.